Functional Comments
Still unsure about this. The only way I was able to get functional comments with a method in the CommentContainer.js is to hold an instance of the whole data object in this.state of CommentContainer to push/concatenate to using the index of involved post. It seems this is incorrect, maybe I misunderstood the directions.