Added Login Functionality
Finished login feature, which allows a user to login and stay logged in. Username is passed down as a prop to the addNewComment method on the comments component so that comments post from entered user. Added a feature on the User Icon in the upper right as an OnClick event which shows a dropdown to the view which shows the username and allows a user to logout